Update To Data
Added a new addedBy field that requires the user creating the bug to add their name in. If I decide to work on this website further, I would add in a way to get the name from the account that is logged in instead of needing to manually input the name.